Refactor ProfilePicture (AvatarImage)
